requirements.txt
streamlit
torch
transformers
Pillow
sentencepiece